refactor(rewrites): fold identical chunk rewrites into a data-driven ChunkRewrite
The chunked_scan, ssd and gated_deltanet rewrites were byte-identical
except for their name/source_type/chunk_type constants, all delegating
to build_chunk_chain. Replace the three classes with a single frozen
ChunkRewrite dataclass parameterised by those constants and register the
three instances from the registry. All four rewrites (chunked_scan, ssd,
gated_deltanet, flash_attention) still register and dispatch by name.
